+++ cmd/cmd.go
@@ -10,6 +10,7 @@ import (
 	"github.com/alwqx/sec/cmd/bond"
 	"github.com/alwqx/sec/cmd/metal"
 	"github.com/alwqx/sec/cmd/quote"
+	"github.com/alwqx/sec/cmd/upgrade"
 	"github.com/alwqx/sec/provider/sina"
 	"github.com/alwqx/sec/types"
 	"github.com/alwqx/sec/version"
@@ -64,6 +65,7 @@ func NewCLI() *cobra.Command {
 		bond.NewBondCLI(), bond.NewBondHistoryCLI(),
 		quote.NewQuoteCLI(), quote.NewQuoteHistoryCLI(),
 		metal.NewMetalCLI(), metal.NewMetalHistoryCLI(),
+		upgrade.NewUpgradeCLI(),
 	)
 
 	return rootCmd